+++ nitrogen/test/src/libraries_test.dart
@@ -1,5 +1,5 @@
 import 'package:code_builder/code_builder.dart';
-import 'package:nitrogen/nitrogen_extension.dart';
+import 'package:nitrogen/src/libraries.dart';
 import 'package:test/test.dart';
 
 void main() {