+++ nitrogen_flutter_svg/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+Export instead of generate extension.
+
 ## 0.2.0+1
 
 * Add dependency on Flutter SDK